old-tests/submission/Microsoft/audio/audio_005.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'HTMLAudioElement' inherits 'playbackRate' from 'HTMLMediaElement' and supports getting its default value</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-playbackrate" />
  <meta name="assert" content="'HTMLAudioElement' inherits 'playbackRate' from 'HTMLMediaElement' and supports getting its default value" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>

    <div id=log></div>
    <audio id="audio0">Your browser does not support media elements.</audio>
    <script type="text/javascript">

    test(function() { assert_equals(document.getElementById("audio0").playbackRate, 1)}, "'HTMLAudioElement' inherits 'playbackRate' from 'HTMLMediaElement' and supports getting its default value");

  </script>
</div>
</body>
</html>
